Add local NFS with container, abstract the PVC
